requirements.txt
Pillow
xlrd
pandas
openpyxl
numpy
google-cloud-vision